they forget about you after 25 turns now of not seeing you

+++ FirstCPPApplication/thinker.h
@@ -21,7 +21,7 @@ class Thinker
 
     public:
         static int visibility_threshold;
-        double turn_last_seen_by_player;
+        unsigned long turn_last_seen_by_player;
         double tracking_memory;
 
         Actor* master;
